requirements.txt
alembic==1.5.8
bcrypt==3.2.0
factory-boy==3.2.0
fastapi==0.63.0
FastAPI-SQLAlchemy==0.2.1
passlib==1.7.4
PyJWT==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.4.0
uvicorn==0.13.4
